moved server response simulator into separate file
